Move some functions from device to verification
